+++ stockanalysis/__init__.py
@@ -1,6 +1,6 @@
 from stockanalysis.stock import Stock
 from stockanalysis.dcf import single_dcf, multi_dcf
-from stockanalysis.utils import create_workbook
+from stockanalysis.sheets import create_workbook, export_dataframe
 from stockanalysis.scrape import scrape_financials, scrape_forecast, scrape_statistics
 
 __all__ = [
@@ -8,6 +8,7 @@ __all__ = [
 	"single_dcf",
 	"multi_dcf",
 	"create_workbook",
+	"export_dataframe",
 	"scrape_financials",
 	"scrape_forecast",
 	"scrape_statistics"
